fix: use id instead of SKU (#39)
Use entity_id in the WHERE clause instead of sku. This adjustment enhances the query's speed, especially when dealing with a large number of products.

During local testing, the export time decreased significantly from 50 minutes to just 8 minutes. Resulting in an increase in speed of 84%. This speed increase only seems to occur if you don't have all numbers as sku's.